Added Decorator to take random moves at the start of a game
